ventas_productos

+++ eliminar_entregado.php
@@ -1,8 +1,13 @@
 <?php 
+include 'functions.php';
 include 'connection.php';
-$id = $_GET['id'];
-$link = connection::link();
-$res = $link->prepare("DELETE FROM ventas WHERE id=$id");
-$res->execute();
-header ('location: tabla_entregados.php');
+if ($_SESSION['rol']==1){
+	$id = $_GET['id'];
+	$link = connection::link();
+	$res = $link->prepare("DELETE FROM ventas WHERE id=$id");
+	$res->execute();
+	header ('location: tabla_entregados.php');
+} else {
+	header ('location:home.php');
+}
 ?>
